Project structure modified, Sign Up & Sign In form component added
