+++ client/src/App.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
+import Dashboard from './components/Dashboard';
 import Home from './components/Home';
-import Login from './components/Login';
 
 function App() {
   return (
     <Router>
       <div>
         <Switch>
-          <Route path="/" exact component={Login} />
-          <Route path="/dashboard" component={Home} />
+          <Route path="/" exact component={Home} />
+          <Route path="/dashboard" component={Dashboard} />
         </Switch>
       </div>
     </Router>
