+++ client/src/components/Dashboard.css
@@ -1,0 +1,24 @@
+.container {
+  width: 90%;
+  max-width: 100rem;
+  height: 100%;
+  margin: 0 auto;
+  position: relative;
+}
+
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 20px;
+  border-bottom: 0.5px solid rgb(155, 155, 155);
+}
+
+.brand {
+  cursor: pointer;
+}
+
+.brand:hover {
+  transform: scale(0.95);
+  transition: transform 0.3s ease-in-out;
+}
