--- client/src/components/Login.css
@@ -1,24 +1,0 @@
-.login-container {
-  width: 40vw;
-  max-width: 50rem;
-  height: 100%;
-  padding: 15px;
-  margin: 20vh auto;
-  position: relative;
-  border: none;
-  box-shadow: 3px 3px 5px 6px #ccc;
-}
-
-.login-form {
-  width: 100%;
-  max-width: 25rem;
-  padding: 15px;
-  margin: auto;
-  height: 100%;
-}
-
-@media only screen and (max-width: 1100px) {
-  .login-container {
-    width: 80vw;
-  }
-}
